+++ apps/core/priv/repo/migrations/20201017153818_create_service_reviews.exs
@@ -6,7 +6,7 @@ defmodule Core.Repo.Migrations.CreateServiceReviews do
       add :id, :uuid, primary_key: true, default: fragment("uuid_generate_v4()"), read_after_writes: true
       add :client_comment, :string, null: true
       add :communication, :integer, null: false
-      add :final_rating, :decimal, null: false
+      add :final_rating, :integer, null: false
       add :pro_response, :string, null: true
       add :professionalism, :integer, null: false
       add :user_id, references(:users, type: :uuid, on_delete: :delete_all), null: true, primary_key: false
